requirements.txt
cryptography>=42.0.0
argon2-cffi>=23.1.0
django-encrypted-model-fields>=0.6.5
django-axes>=6.3.0
django-cors-headers>=4.3.0
django-ratelimit>=4.1.0
django-csp>=3.8
python-decouple>=3.8
django-auditlog>=3.0.0
bandit>=1.7.0
safety>=3.0.0
# Hafta 2
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
djangorestframework-api-key>=3.0.0
